web: refactor SidebarItemView into a separate file
